ripped out all existing css/styles in order to do a complete overhaul
