+++ src/App/App.css
@@ -1,0 +1,16 @@
+/* ───────────────────────────────
+   Palette Reference
+─────────────────────────────── */
+/*
+Theme Colors
+- #e5c5ba hsl(15, 45%, 81%)
+- #3dc091
+- #8d4f5e
+- #1e131c
+- #d130b3 hsl(311, 64%, 50%)
+- #1f6c4a
+- #513f6a
+- #9f21a3
+- #5e2740
+- #431448
+*/
