--- src/App/App.module.css
@@ -1,134 +1,0 @@
-html,
-body {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-  height: 100%;
-  overflow: hidden;
-}
-
-* {
-  box-sizing: inherit;
-}
-
-#root {
-  display: flex;
-  flex-direction: column;
-  height: 100vh;
-  overflow: hidden;
-}
-
-
-body {
-  background: url('/jammingBackground.jpg') center/cover no-repeat, #431448;
-  color: #f9f9f9;
-  font-family: 'Rubik', sans-serif;
-  font-size: 18px;
-}
-
-a {
-  font-weight: 500;
-  color: #5e2740;
-  background-color: #f9f9f9;
-  border-radius: 10px 100px / 120px;
-  opacity: 0.9;
-  padding: 1%;
-  text-decoration: none;
-  margin: 1.5% 0 0 1.5%;
-  display: inline-flex;
-  flex-wrap: wrap;
-}
-
-a:hover {
-  color: #f9f9f9;
-  background-color: hsl(311, 64%, 50%, .6);
-}
-
-h1 {
-  background-color: #010c3f;
-  text-align: center;
-  font-size: 1.88rem;
-  font-size: 3.2em;
-  line-height: 1.1;
-  color: #f9f9f9;
-  text-shadow:
-    3px 3px 0 #431448,
-    -3px 3px 0 #431448,
-    -3px -3px 0 #431448,
-    3px -3px 0 #431448;
-  border-radius: 10px 100px / 120px;
-  letter-spacing: 0.2rem;
-}
-
-h1 .highlight {
-  color: #3dc091;
-}
-
-.header,
-.footer {
-  color: #e5c5ba;
-  text-shadow:
-    1px 1px 0 #431448,
-    -1px 1px 0 #431448,
-    -1px -1px 0 #431448,
-    1px -1px 0 #431448;
-  position: absolute;
-  background-color: hsl(311, 64%, 50%, .5);
-}
-
-footer {
-  position: fixed;
-  left: 1%;
-  bottom: 1%;
-  margin: 7px;
-  color: #513f6a !important;
-  letter-spacing: 0.2rem;
-}
-
-@media (max-: 690px) {
-
-  .header-links {
-    position: fixed;
-    right: 2%;
-    top: 0.5%;
-    z-index: 10;
-    display: flex;
-    flex-direction: column;
-  }
-
-  .header-link {
-    display: flex;
-    font-size: smaller;
-    flex-wrap: wrap;
-    padding: 1%;
-  }
-}
-
-
-@media only screen and (max-width: 668px) {
-  body {
-    background-image: url('/jammingSM.jpg');
-    background-size: cover;
-    background-position: center;
-    font-size: 16px;
-  }
-}
-
-
-/*
-https://imagecolorpicker.com
-Theme Colors
-
-- #e5c5ba hsl(15, 45%, 81%)
-- #3dc091
-- #8d4f5e
-- #1e131c
-- #d130b3 hsl(311, 64%, 50%)
-- #1f6c4a
-- #513f6a
-- #9f21a3
-- #5e2740
-- #431448
-
-https://htmlcolors.com/hex-to-hsl
-*/
