+++ src/components/Button.tsx
@@ -1,8 +1,6 @@
-import styles from "./Button.module.css"
-
 function Button() {
     return (
-        <div className={styles.Button}>
+        <div>
             {/* <button class="Track-action"><!-- + or - will go here --></button> */}
         </div>
     );
