+++ src/components/Tracklist.tsx
@@ -1,5 +1,4 @@
-import Track from "../track/Track";
-import styles from "./Tracklist.module.css"
+import Track from './Track';
 
 export type Track = {
     name: string;
@@ -14,7 +13,7 @@ export type Track = {
 
 function Tracklist(props: { userSearchResults: Track[]; isRemoval: boolean; onAdd: (track: Track) => void; onRemove: (track: Track) => void; onArtistClick: (artistId: string) => void; isArtistClickable?: boolean; }) {
     return (
-        <div className={styles.TrackList}>
+        <div>
 
             {props.userSearchResults.length > 0 ? (props.userSearchResults?.map((track) => (
                 <Track track={track} key={track.id} isRemoval={props.isRemoval} onAdd={props.onAdd} onRemove={props.onRemove} onArtistClick={props.onArtistClick} isArtistClickable={props.isArtistClickable} />
