--- src/components/playlist/Playlist.module.css
@@ -1,49 +1,0 @@
-.Playlist {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    overflow-y: scroll;
-    max-height: 50vh;
-    padding: 2rem;
-    background-color: rgba(1, 12, 63, 0.7);
-    box-shadow: 0 4px 2px 2px #000000;
-    margin-top: -85px;
-}
-
-.Playlist input {
-    border: 0;
-    outline: 0;
-    background: transparent;
-    border-bottom: 1px solid #6f6f6f;
-    font-family: 'Poppins', sans-serif;
-    font-size: 1.5rem;
-    color: #fff;
-    width: 100%;
-    font-weight: 900;
-}
-
-.Playlist-save {
-    cursor: pointer;
-    padding: .77rem;
-    border-radius: 54px;
-    border-width: 0px;
-    margin-top: 1.27rem;
-    background-color: #6c41ec;
-    text-align: center;
-    font-size: .83rem;
-    transition: background-color .25s;
-    color: #fff;
-    font-weight: 500;
-}
-
-.Playlist-save:hover {
-    background-color: rgba(108, 65, 233, .7);
-}
-
-::-webkit-scrollbar {
-    background: transparent;
-}
-
-.track-artist:hover {
-    text-decoration: none;
-}
