--- src/components/searchbar/Searchbar.module.css
@@ -1,41 +1,0 @@
-.SearchBar {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    padding-top: 6.94rem;
-    margin-bottom: 6.33rem;
-}
-
-.SearchBar input {
-    padding: .88rem 0;
-    border: 1px solid #fff;
-    border-radius: 3px;
-    margin-bottom: 2.22rem;
-    color: #010c3f;
-    text-align: center;
-    font-size: 1rem;
-    width: 300px;
-}
-
-.SearchBar input:focus {
-    outline: none;
-}
-
-.SearchButton {
-    cursor: pointer;
-    border-radius: 54px;
-    text-align: center;
-    font-size: .833rem;
-    transition: background-color .25s;
-    border: 0px;
-    font-weight: 500;
-    width: 100px;
-}
-
-.SearchButton:hover {
-    background-color: rgba(108, 65, 233, .7);
-}
-
-.TypeButton:hover {
-    background-color: rgba(108, 65, 233, .7);
-}
