--- src/components/track/Track.module.css
@@ -1,23 +1,0 @@
-.Track {
-    display: flex;
-    align-items: center;
-    border-bottom: 1px solid rgba(256, 256, 256, 0.8);
-}
-
-.Track-action {
-    cursor: pointer;
-    padding: .5rem;
-    font-size: 1.05rem;
-    transition: color .25s;
-    border: 0px;
-    background-color: rgba(0, 0, 0, 0);
-    color: #fff;
-}
-
-.Track-action:hover {
-    color: rgba(265, 265, 265, .7);
-}
-
-.track-artist:hover {
-    text-decoration: underline;
-}
